libbpf-tools/profile: Exclude use of Kconfig for PAGE_OFFSET
page_offset can be computed by Kconfig at runtime.
However, if Kconfig is not in the target, libbpf will stop loading the program
with an error.
So instead of reading the Kconfig, it uses the magic number 0xff000000000000000
to check whether kernel_ip is the kernel address or not. This is the value
recommended by the ebd team. It works for webOS TV and webOS OSE rpi-64, but not
sure about other targets.

Change-Id: Iaa2eb35d57c0f11e5ca0e2fff8ce1a7ec487e835