libbpf-tools/memleak: Expand value size to 64bit
Since the value size was set to fit 32-bit, it was not possible to
create a stack_trace map on a TV platform with a 64-bit kernel. So,
change the value size to 64-bit size to make it work in the multilib
environment. And the size of stack_storage _size was too small.
Therefore, an error occurs when a value of a larger size is required, so
the size is increased.

Change-Id: I1582f9b15d28b675905175f83e495b4b5c5aa7a6